Sound: Disable analogue sound when the Pi Hat is attached
The Pi Hat's LEDs are driven with the RPi's PWM pin which is shared with
the audio jack, meaning that there is interference with the audio when
the LEDs are programmed and vice versa. Resolve this problem by
disabling the snd_bcm2835 kernel sound module when the Pi Hat is
attached.