Pi Hat: Add DBus signal for power button press
Register a DBus signal in the Pi Hat high-level service for when the
power button is pressed to allow other applications to trigger on it.
This also circumvents the problem with multiple processes trying to lock
access to the library when they require events triggered on the power
button press.