Disabled SpeakerLEDsService due to PiHat and PowerHat lib
The Speaker LED board uses the i2c GPIO pin for detection. The same
pin is used for the power button on the PiHat and PowerHat libs which
lock the usage of it. This automatically breaks the detection of the
board in the current state. Disabling the service until we provide
support for it again.