Disable LED Speaker and add PowerHat detection
The SpeakerLedService was disabled previously in kano-boards-daemon
due to detection issues and drop of support. By disabling it in the
animation routines, we no longer retry for 5s to grab the DBus iface.

kano-speakerleds has a command for detection. Disabled the LED Speaker
from here and also added the missing Power Hat dectection. Return
codes were added as needed.