Fixed an issue where animations would not unlock the service API
The D-Bus service API for turning on the LEDs was only unlocked
in the signal handler used to stop animations cleanly, but was
not called when the animation itself finished as expected.
Consequently, the service would take longer to unlock it itself
and animations below the lock would take longer to reappear and
new animations would either fail after the retries or take just
as long to start again.