Architecture refactor, pi_hat lib fixes, board hotplugging, & more
This is changeset consists of the several parts:

1. Architecture refactor

Previously, the ServiceManager started all services for all boards.
This meant that the pi_hat library locked GPIO pins for PiHat and
PowerHat boards which broke the service for the LED Speaker.
Instead, there is now a DeviceDiscovery service which runs detection
routines for each peripheral and only brings a service up once a
device was detected. It is also a cleaner interface as only the
usable APIs are exposed for devices that are already plugged in.

2. pi_hat library fixes

There were a few issues with the callback mechanism used to get
notifications from the hardware (ISRs). Some of these included
not being able to register more than one, pointers to more than
one were garbage collected by Python, callback lists where reset
when registering a new cb, and detection signatures needed 0s.

3. Board hotplugging feature

With the DiscoveryService and device services coming online only
when peripherals are plugged in meant that using DBus signals
for disconnecting and connecting events we could support
plugging and unplugging boards at runtime.

4. Support for LED Speaker

The architecture refactor enabled the animation support for the
LED Speaker again given that the PiHat and PowerHat boards were
not locking the GPIO pins anymore.

5. Some slight performance improvements

Particularly at daemon startup, the detection routine should now
start quicker with tweaks to the GOBject detection threads and
not waiting for services to be started on the bus.