Starting PiHat and PowerHat services with power button enabled
This is moving away from relying on a bootup script from the repo
kano-desktop which enabled the power button at boot. That script
was removed and the builtin timer within the button callback
effectively does the same thing. This is also in support for
board hotplugging as the service will now start with the button
enabled and will work thereafter.