CPU monitor animation robustness with hotplugging
The animation now has a reconnect function that helps to account
for when boards get unplugged and plugged in live. When this
happens, the iface object will not reach the DBus service and will
raise an exception at which point we go into the reconnect loop.
This might be improved by using DBus signal handlers for when the
boards get plugged in rather than polling for the service. TODO.