Fix audio driver initialisation clash
The issue here was to do with when KanoHatLeds object was created
in relation to when the snd_bcm2835 module was loaded. The object
cannot be created after the module is loaded, only before. As such,
moved the module loading after creating the object as early as
possible. This takes into account turn-off-hat-leds.service at
boot up. Bottom line, the audio module can only be loaded as soon
as kano-boards-daemon is started.