Added functionality to detect CKCs, CKC v1, CKC v1.1
The additional CKC v1 and v1.1 come from different manufacturing
batches that we want to tell apart. Specifically, the v1 and v1.1
only differ in terms of screen model name in EDID: MST-HDMI in v1
and HTC-HDMI in v1.1. Additionally, given that there are times when
these functions are required to be run outside the lifecycle of the
kano-boards-daemon, an option was added to directly run the detection
routine directly through their respective libraries (with_dbus).