Hidden RC for turn-off-hat-leds service, fire and forget unit
Instead of reporting a non-zero return code back to systemd, always
report zero and do not show systemd failure messages to the user at
bootup. If there is a tbk raised, it is logged to the journal via
a simple print. For context, this script seemed to fail when there
is low-voltage on the Pi, i.e. CKC + battery.