CKC & CKT: Ignore hat presence when discerning kit version
The original implementation of the functions to check if the CKC and CKT
were used involved checking for the presence of the PowerHat and then
using the EDID to discern between the two. Instead, drop the hat check
for the CKC as we have all the EDIDs that will ever be used for those
kits so can reliably query the kit status. Migrate the CKT's
implementation to require the PowerHat but only require that, in
addition, it isn't a CKC for safety.