Ensure snd_bcm2835 audio module gets loaded even during a crash
There is a chance that the kano-boards-daemon fails and the module
never gets loaded. Trying the OS on a RPi 3B+, libkano-hat causes
the daemon to crash. Wrap the lib to handle a potential exception
and also load the module at the top level in the daemon if it fails.